refactor: _components layout, pages 디렉토리 분리
