+++ apps/frontend/src/_components/pages/result/ButtonGroupSection.tsx
