+++ apps/frontend/src/_components/pages/result/ProgressBar.tsx
