+++ apps/frontend/src/_components/pages/result/ScoreSection.module.css
