+++ apps/frontend/src/_components/pages/result/ScoreSection.tsx
