+++ apps/frontend/src/_components/pages/result/TotalEvaluationSection.module.css
