+++ apps/frontend/src/_components/pages/result/TotalEvaluationSection.tsx
