+++ apps/frontend/src/app/result/page.tsx
@@ -8,10 +8,10 @@ import { useQuery } from "@tanstack/react-query";
 import { selectChatId } from "@/_store/redux/features/chat/selector";
 import { useInterviewerStore } from "@/_store/useInterviewerStore";
 import Text from "@repo/ui/Text";
-import ScoreSection from "@/_components/result/ScoreSection";
-import TotalEvaluationSection from "@/_components/result/TotalEvaluationSection";
-import QuestionEvaluationSection from "@/_components/result/QuestionEvaluationSection";
-import ButtonGroupSection from "@/_components/result/ButtonGroupSection";
+import ScoreSection from "@/_components/pages/result/ScoreSection";
+import TotalEvaluationSection from "@/_components/pages/result/TotalEvaluationSection";
+import QuestionEvaluationSection from "@/_components/pages/result/QuestionEvaluationSection";
+import ButtonGroupSection from "@/_components/pages/result/ButtonGroupSection";
 
 interface Scores {
   standard: string;
